+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Version 1.22 (current master)
 * enhanced support to get versioned element from MongoDB DAOs
+* improved and reworked GenericConverter
 * adds an option to send messages to an alternative topic via IVORefreshSender
 
 # Version 1.21